Read mod status differently
When user only posts emote, emoteonly=1 is added to line and that messes
up mod status detection. Now we only look for mod=0 and mod=1 in line.
This also means that if mod has mod=0 in his chat message he will be
detected as normal user instead of mod... but I think that doesn't
matter as that is not something that can happen accidentally.